Can now draw new line based on incoming data.
This hasn't been tested with more than a single new point being added, but I have no
reason to beleive that it wouldn't work.

It reads out the old features, then appends the new points to the old geometry,
creating a new vector.feature, appending it to the map, and deleting the old one.
As a side effect, this would make it really easy to simplify the line each time
new points are added.