Much cleaner way to interact with data.
Now points hold the important data, and they're simply added on to.  The
line is regenerated each time, but we only have to add to the list of
clickable points.  This will make it MUCH easier to insert points out-of-order.

Now, just need to add reordering of the list each time the linestring is regenerated.
Could also refactor things for code reuse, but it isn't a big deal since it's
only repeated in the init function.